worker/requirements.txt
# Core dependencies
slack-bolt
requests
boto3
bedrock-agentcore==1.2.0

# MCP and Strands dependencies
mcp>=0.9.0
strands-agents==1.23.0
strands-agents-tools==0.2.14

# AWS OpenTelemetry Distro - REQUIRED for AgentCore observability
aws-opentelemetry-distro>=0.10.0

# Additional dependencies
pydantic
pydantic-core
rpds-py
annotated-types
typing-extensions